Add AppCompat presenter for Android
