+++ MvvmCross-AndroidSupport/MvvmCross.Droid.Support.V7.AppCompat/MvvmCross.Droid.Support.V7.AppCompat.csproj
@@ -99,6 +99,7 @@
     <Compile Include="Widget\MvxAppCompatRadioGroup.cs" />
     <Compile Include="Widget\MvxAppCompatSpinner.cs" />
     <Compile Include="MvxAppCompatSetup.cs" />
+    <Compile Include="MvxAppCompatViewPresenter.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
