+++ TestProjects/Navigation/RoutingExample.Droid/Setup.cs
@@ -10,6 +10,7 @@ using MvvmCross.Droid.Views.Attributes;
 using RoutingExample.Core;
 using System;
 using System.Threading.Tasks;
+using MvvmCross.Droid.Support.V7.AppCompat;
 
 namespace RoutingExample.Droid
 {
@@ -26,45 +27,7 @@ namespace RoutingExample.Droid
 
         protected override IMvxAndroidViewPresenter CreateViewPresenter()
         {
-            return new MyViewPresenter(AndroidViewAssemblies);
-        }
-    }
-
-    public class MyViewPresenter : MvxAndroidViewPresenter
-    {
-        public MyViewPresenter(IEnumerable<Assembly> androidViewAssemblies) : base(androidViewAssemblies)
-        {
-        }
-
-        protected new AppCompatActivity CurrentActivity => (AppCompatActivity)base.CurrentActivity;
-        protected new FragmentManager CurrentFragmentManager => CurrentActivity.SupportFragmentManager;
-
-        protected override void ShowDialogFragment(Type view, 
-           MvxDialogAttribute attribute, 
-           MvxViewModelRequest request)
-        {
-            var dialog = Fragment.Instantiate(CurrentActivity, FragmentJavaName(attribute.ViewType)) as DialogFragment;
-            dialog.Show(CurrentFragmentManager, attribute.ViewType.Name);
-        }
-
-        protected override IMvxFragmentView CreateFragment(string fragmentName)
-        {
-            var fragment = Fragment.Instantiate(CurrentActivity, fragmentName);
-            return fragment as IMvxFragmentView;
-        }
-
-        protected override void ShowFragment(Type view,
-            MvxFragmentAttribute attribute,
-            MvxViewModelRequest request)
-        {
-            ShowHostActivity(attribute);
-
-            var fragmentName = FragmentJavaName(attribute.ViewType);
-            var fragment = CreateFragment(fragmentName);
-
-            var ft = CurrentFragmentManager.BeginTransaction();
-            ft.Replace(attribute.FragmentContentId, fragment as Fragment, fragmentName);
-            ft.CommitNowAllowingStateLoss();
+            return new MvxAppCompatViewPresenter(AndroidViewAssemblies);
         }
     }
 }
